Extract discoverOutboundIP to standalone file
runner_start.go is gated behind //go:build linux, which made
discoverOutboundIP unavailable on darwin. Since the function is
pure net package with nothing linux-specific, moving it to its
own file without a build constraint fixes the cross-platform
build.
